requirements.txt
streamlit
requests
python-dotenv
streamlit_js_eval